Fix editable grid save operation
The save operation on editable grid was not optimal. It was
sending all records and fetching all fields of those records.

With this fix, now editable grid will send only changed record
to save and will fetch only fields included in grid.

Fixes RM-10330

+++ axelor-web/src/main/webapp/js/widget/widget.slickgrid.js
@@ -1826,7 +1826,14 @@ Grid.prototype.__saveChanges = function(args, callback) {
 		});
 	}
 
-	return saveDS.saveAll(records).success(function(records, page) {
+	records = _.where(records, { $dirty: true });
+	if (records.length === 0) {
+		return setTimeout(focus, 200);
+	}
+
+	var fields = handler.selectFields ? handler.selectFields() : undefined;
+
+	return saveDS.saveAll(records, fields).success(function(records, page) {
 		if (data.getItemById(0)) {
 			data.deleteItem(0);
 		}
